fix np.float issue

+++ allesfitter/basement.py
@@ -288,7 +288,7 @@ class Basement():
             self.settings['fast_fit'] = False
         
         if ('fast_fit_width' in self.settings.keys()) and len(self.settings['fast_fit_width']):
-            self.settings['fast_fit_width'] = np.float(self.settings['fast_fit_width'])
+            self.settings['fast_fit_width'] = np.float64(self.settings['fast_fit_width'])
         else:
             self.settings['fast_fit_width'] = 8./24.
                 
@@ -562,10 +562,10 @@ class Basement():
                 t_exp = self.settings['t_exp_'+inst].split(' ')
                 #if float
                 if len(t_exp)==1:
-                    self.settings['t_exp_'+inst] = np.float(t_exp[0])
+                    self.settings['t_exp_'+inst] = np.float64(t_exp[0])
                 #if array
                 else:
-                    self.settings['t_exp_'+inst] = np.array([ np.float(t) for t in t_exp ])
+                    self.settings['t_exp_'+inst] = np.array([ np.float64(t) for t in t_exp ])
             #::: if not given / given as an empty field
             else:
                 self.settings['t_exp_'+inst] = None
@@ -673,7 +673,7 @@ class Basement():
         for i,key in enumerate(self.allkeys):
             #::: if it's not a "coupled parameter", then use the given value
             if np.atleast_1d(buf['value'])[i] not in list(self.allkeys):
-                self.params[key] = np.float(np.atleast_1d(buf['value'])[i])
+                self.params[key] = np.float64(np.atleast_1d(buf['value'])[i])
             #::: if it's a "coupled parameter", then write the string of the key it is coupled to
             else:
                 self.params[key] = np.atleast_1d(buf['value'])[i]
@@ -909,9 +909,9 @@ class Basement():
         self.bounds = [ str(item).split(' ') for item in buf['bounds'][ self.ind_fit ] ] #len(ndim)
         for i, item in enumerate(self.bounds):
             if item[0] in ['uniform', 'normal']:
-                self.bounds[i] = [ item[0], np.float(item[1]), np.float(item[2]) ]
+                self.bounds[i] = [ item[0], np.float64(item[1]), np.float64(item[2]) ]
             elif item[0] in ['trunc_normal']:
-                self.bounds[i] = [ item[0], np.float(item[1]), np.float(item[2]), np.float(item[3]), np.float(item[4]) ]
+                self.bounds[i] = [ item[0], np.float64(item[1]), np.float64(item[2]), np.float64(item[3]), np.float64(item[4]) ]
             else:
                 raise ValueError('Bounds have to be "uniform", "normal" or "trunc_normal". Input from "params.csv" was "'+self.bounds[i][0]+'".')
